feature: refactor flink and flinkful dependency
