+++ pom.xml
@@ -138,9 +138,6 @@
         <postgresql.version>42.3.3</postgresql.version>
         <druid.version>0.22.1</druid.version>
         <calcite-druid.version>1.29.0</calcite-druid.version>
-        <flink.version>1.13.6</flink.version>
-        <flink-kubernetes-operator.version>1.1.0</flink-kubernetes-operator.version>
-        <flinkful.version>1.13.6_4-SNAPSHOT</flinkful.version>
         <config.version>1.3.3</config.version>
         <okhttp.version>4.10.0</okhttp.version>
         <guava.version>31.0-jre</guava.version>
@@ -148,8 +145,8 @@
         <milky.version>1.0.7</milky.version>
         <hadoop.version>3.3.4</hadoop.version>
         <akka.version>2.5.21</akka.version>
-        <sofa.ark.version>2.0.0</sofa.ark.version>
         <protobuf.version>3.21.5</protobuf.version>
+        <netty.version>4.1.82.Final</netty.version>
     </properties>
 
     <dependencyManagement>
@@ -268,6 +265,11 @@
             </dependency>
             <dependency>
                 <groupId>${project.groupId}</groupId>
+                <artifactId>scaleph-plugin-flink-1.13</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${project.groupId}</groupId>
                 <artifactId>scaleph-core-di</artifactId>
                 <version>${project.version}</version>
             </dependency>
@@ -315,22 +317,6 @@
                 <groupId>io.swagger</groupId>
                 <artifactId>swagger-annotations</artifactId>
                 <version>${swagger.version}</version>
-            </dependency>
-
-            <dependency>
-                <groupId>com.alipay.sofa</groupId>
-                <artifactId>sofa-ark-springboot-starter</artifactId>
-                <version>${sofa.ark.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.alipay.sofa</groupId>
-                <artifactId>sofa-ark-all</artifactId>
-                <version>${sofa.ark.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.alipay.sofa</groupId>
-                <artifactId>sofa-ark-api</artifactId>
-                <version>${sofa.ark.version}</version>
             </dependency>
 
             <dependency>
@@ -415,68 +401,6 @@
                 <version>${guava.version}</version>
             </dependency>
 
-            <!-- flink -->
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-java</artifactId>
-                <version>${flink.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-streaming-java_${scala.binary.version}</artifactId>
-                <version>${flink.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-table-planner_${scala.binary.version}</artifactId>
-                <version>${flink.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-table-planner-blink_${scala.binary.version}</artifactId>
-                <version>${flink.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-runtime-web_${scala.binary.version}</artifactId>
-                <version>${flink.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-clients_${scala.binary.version}</artifactId>
-                <version>${flink.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-yarn_${scala.binary.version}</artifactId>
-                <version>${flink.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-kubernetes_${scala.binary.version}</artifactId>
-                <version>${flink.version}</version>
-            </dependency>
-
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-kubernetes-operator</artifactId>
-                <version>${flink-kubernetes-operator.version}</version>
-                <exclusions>
-                    <exclusion>
-                        <groupId>org.apache.flink</groupId>
-                        <artifactId>flink-clients</artifactId>
-                    </exclusion>
-                    <exclusion>
-                        <groupId>org.apache.flink</groupId>
-                        <artifactId>flink-runtime-web</artifactId>
-                    </exclusion>
-                    <exclusion>
-                        <groupId>org.apache.logging.log4j</groupId>
-                        <artifactId>log4j-slf4j-impl</artifactId>
-                    </exclusion>
-                </exclusions>
-            </dependency>
-
             <dependency>
                 <groupId>org.apache.hadoop</groupId>
                 <artifactId>hadoop-common</artifactId>
@@ -548,38 +472,6 @@
                 <groupId>com.typesafe</groupId>
                 <artifactId>config</artifactId>
                 <version>${config.version}</version>
-            </dependency>
-
-            <dependency>
-                <groupId>cn.sliew</groupId>
-                <artifactId>flinkful-cli-base</artifactId>
-                <version>${flinkful.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>cn.sliew</groupId>
-                <artifactId>flinkful-cli-descriptor</artifactId>
-                <version>${flinkful.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>cn.sliew</groupId>
-                <artifactId>flinkful-cli-frontend</artifactId>
-                <version>${flinkful.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>cn.sliew</groupId>
-                <artifactId>flinkful-rest-client</artifactId>
-                <version>${flinkful.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>cn.sliew</groupId>
-                <artifactId>flinkful-kubernetes-operator</artifactId>
-                <version>${flinkful.version}</version>
-                <exclusions>
-                    <exclusion>
-                        <groupId>org.apache.flink</groupId>
-                        <artifactId>flink-kubernetes-operator</artifactId>
-                    </exclusion>
-                </exclusions>
             </dependency>
 
             <dependency>
@@ -655,8 +547,7 @@
             <dependency>
                 <groupId>io.netty</groupId>
                 <artifactId>netty-all</artifactId>
-                <!--                <version>${netty.version}</version>-->
-                <version>4.1.82.Final</version>
+                <version>${netty.version}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
