+++ scaleph-engine/scaleph-engine-seatunnel/pom.xml
@@ -45,28 +45,9 @@
             <groupId>${project.parent.groupId}</groupId>
             <artifactId>scaleph-plugin-seatunnel-connectors</artifactId>
         </dependency>
-
         <dependency>
-            <groupId>cn.sliew</groupId>
-            <artifactId>flinkful-cli-base</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>cn.sliew</groupId>
-            <artifactId>flinkful-cli-descriptor</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>cn.sliew</groupId>
-            <artifactId>flinkful-rest-client</artifactId>
-        </dependency>
-
-        <!-- seatunnel additional dependency -->
-        <dependency>
-            <groupId>org.apache.flink</groupId>
-            <artifactId>flink-table-planner_${scala.binary.version}</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.flink</groupId>
-            <artifactId>flink-table-planner-blink_${scala.binary.version}</artifactId>
+            <groupId>${project.parent.groupId}</groupId>
+            <artifactId>scaleph-plugin-flink-1.13</artifactId>
         </dependency>
 
         <dependency>
