+++ scaleph-plugins/scaleph-plugin-flink/scaleph-plugin-flink-1.13/pom.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>cn.sliew</groupId>
+        <artifactId>scaleph-plugin-flink</artifactId>
+        <version>1.0.2-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+    <artifactId>scaleph-plugin-flink-1.13</artifactId>
+
+    <properties>
+        <flink.version>1.13.6</flink.version>
+        <flink-kubernetes-operator.version>1.1.0</flink-kubernetes-operator.version>
+        <flinkful.version>1.13.6_4-SNAPSHOT</flinkful.version>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>${project.parent.groupId}</groupId>
+            <artifactId>scaleph-plugin-flink-api</artifactId>
+        </dependency>
+
+        <!-- flinkful -->
+        <dependency>
+            <groupId>cn.sliew</groupId>
+            <artifactId>flinkful-cli-base</artifactId>
+            <version>${flinkful.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>cn.sliew</groupId>
+            <artifactId>flinkful-cli-descriptor</artifactId>
+            <version>${flinkful.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>cn.sliew</groupId>
+            <artifactId>flinkful-rest-client</artifactId>
+            <version>${flinkful.version}</version>
+        </dependency>
+
+        <!-- seatunnel additional dependency -->
+        <dependency>
+            <groupId>org.apache.flink</groupId>
+            <artifactId>flink-table-planner_${scala.binary.version}</artifactId>
+            <version>${flink.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.flink</groupId>
+            <artifactId>flink-table-planner-blink_${scala.binary.version}</artifactId>
+            <version>${flink.version}</version>
+        </dependency>
+    </dependencies>
+
+</project>
