+++ scaleph-plugins/scaleph-plugin-flink/scaleph-plugin-flink-api/src/main/java/cn/sliew/scaleph/plugin/flink/FlinkfulClientProvider.java
@@ -16,20 +16,8 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.engine.seatunnel.service.util;
+package cn.sliew.scaleph.plugin.flink;
 
-public enum QuartzJobUtil {
-    ;
-
-    public static String getJobName(String projectCode, Long jobCode) {
-        return String.format("%s_%s", projectCode, jobCode);
-    }
-
-    public static String getFlinkBatchJobName(String jobName) {
-        return "FLINK_BATCH_JOB_" + jobName;
-    }
-
-    public static String getFlinkBatchTriggerKey(String jobName) {
-        return "FLINK_BATCH_TRI_" + jobName;
-    }
+public interface FlinkfulClientProvider {
+    
 }
